Add redux "Nothing is Missable" option.
Currently this option adds a button to Kanalet Castle upstairs to raise the gate, and makes it so you can get all seashells from mansion no matter how many you have. The latter ONLY works on new save files and is not retroactive. More features are planned for this option.