More bed bugs: add pathing blocker to top of bed to prevent sprite drifting upward.
This has been an ongoing series at this point.
https://github.com/BigheadSMZ/Links-Awakening-DX-HD/commit/567f486a6b04a5fb5436ef0e31a09f4cf19b88da
https://github.com/BigheadSMZ/Zelda-LA-DX-HD-Updated/commit/eb96898954562f84c1117a22435467757fe44ed7
I have probably jumped into this bed almost 400 times trying to find bugs, find ways to fix it, and still keep finding new ways to break it. I am hoping this is finally the last one. You have to REALLY be trying to break it to find something now.