Always play correct overworld music after Marin photograph on beach cliff.
The music after the photo was set to just repeat whatever played last. If the last song that was started was the overworld music with the intro, then it would play that again. Because no other music will play during this point, just always start the "introless" version of overworld music.