Fix Marin beach photo music transition without breaking switching to other tracks.
Last version of the fix broke the transition going into Mabe Village. I should always test just a little bit more... The problem was the hack to not restart the music after getting the sword, it was breaking this situation not allowing to swap the track and instead just running "return". When Marin is with the player at this point in the game,"maria_state" will be set to "3" so that check should suffice. I don't think I'll need to check state "8" for the later "walk" but I'll keep the photo option open just in case.