Hardhit Beetle: Fix text spam when returning to blue & move projectile to bottom draw layer.
This boss had a bug that when hit for the first time, it enabled a 2 second loop that checked if it turned back to blue and if so, spit out a text message. The bug was that it would continue to show text messages even if it remained blue, so every two seconds it was blue after the first hit a text message would pop up. The second "fix" moves the projectile to a lower draw layer so it doesn't overlap Link's sprite. This is not perfect either, but it looks less awkward this way.