Modify colored version of Ghost photo to better match in-game sprite colors.
This always bugged me. Soon as the photo is snapped, everything changed to a different color. The yellow grass turned green, the ghost became purple, the trees got lighter, the flowers more detailed. It just looked weird. So I recolored it to better match the sprites that set up the photo.