Don't store save path in `settings` and use `Path.Combine` over concatenation.
Having the save file path in the save file's was kind of counter intuitive. If the user decides to move from or to a "portable.txt" simply moving the save files still won't load them. The settings file needed to be edited with the new save path which kind of defeats the purpose. So save files are now locked to "AppData\Zelda_LA" or the base game folder if "portable.txt" is created with no nuance. Also, replace a bunch of concatenated paths with `Path.Combine` so we don't need to worry about the proper use of slashes around the codebase.