Fix actions that make "Blocking" state get stuck causing sound effect to be silent.
Sometimes when drawing the shield, the sound effect won't play. This is because the "_blockButton" field which must be false for the sound effect to play can get stuck when performing certain actions and the shield is put away. For instance, when blocking and then picking up an item, or talking to an NPC, these stop Link from blocking but never reset the "_blockButton" state.