Fix exiting dungeon 8 underground passage to unspawned staircase.
At one of the rightmost rooms of dungeon 8 above the miniboss that throws the ball at you, is a staircase (d8_2d_6R) that takes you to a room that has a "secret" staircase (d8_2d_6L) that only spawns when defeating the enemies in the room. If you exit the underground passage into that room before the stairs are spawned, you will end up at position 0,0 in the middle of nowhere because the game doesn't know where to put you. So to circumvent this, just spawn the stairs (d8_2d_6L) when entering the opposite staircase (d8_2d_6R).