Delete a bunch of junk/unused/debug/stale/worthless code.
None of this is useful to me or probably anyone. Most of it is code that was replaced but left commented off. I also don't debug issues in the same way as the original developers so all of these leftover debug tools are useless to me. It's a far cry from "cleaning up" the codebase, but it's a start.