Don't continue playing music after instrument collection and screen goes white.
Not exactly sure when this started happening, but it's now fixed. If I had to guess, it was the change that played the "dungeon clear" music throughout the entire dungeon after collecting the heart and until the instrument is collected. The change should only intercept the dungeon music with the cleared music and not change the logic but who knows.