FreeCam: Right stick adjusts camera center; pressing right stick resets center to Link.
This is the "Free Camera" option that has been requested for some time. It works well enough and feels natural to the point that no special options are required to use it. Just press on the right stick in a direction to move the camera, press in the stick to reset the camera. Works with both normal camera or smooth camera.