Patcher: Add a workaround to language files in Backup folder.
While this should be a "temporary" solution, there is no choice now but to keep it forever now, as there is no telling if the user will have old language files in their backup folder. The root probably should be fixed, but this will always fix it anyway, so it's whatever I guess. The hash for "newHash" was zeroed out so it can be applied to v1.3.0 without it saying "already patched". This is an artificial limitation anyway, since any version can patch any version due to using backup files.