Greatly simplify the dungeon barrier jump-fix hack.
Now that I have a better grasp on things, lets do this in a cleaner way. It's still not as "smooth" as I'd like it to be, as a hack is a hack. The proper way to fix this would be to somehow catch it in "SystemBody.cs" in the "UpdateVelocityZ" method. But that is far more complex since it governs all bodies and not just Link's. Another day perhaps.