Remove dungeon barrier when pushing block over top of it (Level 7).
There is a block that can be pushed over a dungeon barrier in level 7 when it's in the "down" state. When the switch is hit again, the barrier can pop though the block which looks really weird, and did not happen in the original game. A user also seen a bug where it popped through when in the down state. Because of the way objects are drawn, there is no easy fix there. So the simple solution: delete the barrier.