Code refactoring, render target failsafes, rename EditorUI to UIManager.
This started as a small cleanup to "Game1" and became something much larger. I once again tried to solve the exclusive fullscreen issue with render targets getting lost and/or going null but to no avail. It truly is annoying and I don't think I'll ever be able to figure it out as I'm currently too stupid to understand how it all works. Exclusive Fullscreen can now be set with a variable in the save file. I also renamed "EditorUI" to the more generic "UIManager" since that is the class it's based off of, and it includes the entire UI, not just the editor UI. So its current name made no sense.