Add another name entry easter egg using name "ZeldaBop".
This plays a remix of the original "The Legend of Zelda" theme. It's another potentially unused song, at least I have no idea where it is used if it is. It sounds like it could have been used for the credits at some point, but was later replaced with the current credits music. Either way, now there's a way to listen to it.