Replace instrument sequence GbsPlayer stop time check with delta time counter.
I'm honestly not sure why the original dev(s) went with the method of setting a "stop time" for the playing music then made a check that relied on the music stopping after approximately 8 seconds to continue the instrument sequence. The rest of the sequence is already using game delta time, so why this in this one instance? I can't even fathom a guess. At most it is about 50-80ms more accurate which is basically irrelevant. I know there's always a reason behind every decision, but no clue what the reason was for this one.