Fix a really dumb decision/mistake that could wipe out the first save file.
Apparently there was no protection put in place if the game did not have a save loaded and it tries to save. It automatically tries to write to save slot 1. My failure was forgetting that the main menu quit button and the in-game quit button share the same "page", so the main menu is not supposed to try to save anything. One tiny little variable was all that existed between proper operation and complete obliteration of a save file...