Rework the way Marin spawns when added as follower after level 8.
This gets rid of the old code and reuses the code for when Marin spawns at the beach. This fixes two key issues. The first issue is that there was a "spawn limit" to how many times you could have Marin follow you before she bugged out and disappeared. This fixed itself after a map change, but not it's fixed permanently. The second fix prevents multiple Marins from spawning and stacking after she returns to her spot by the weathervane. This wasn't apparent before, but now that her animation was fixed it was much easier to notice multiple Marins.