Add more sound effects to menus and fix distortion in "menu close" sound.
This adds sound effects to the game menus. When going "back" a menu, it plays the same sound as when closing the in-game pause menu. When moving a slider, it plays the standard "beep" as when changing options or inventory selection. This also fixes distortion (or a popping sound) at the beginning of the sound effect that is played when the menu is closed.