Fix 77 sound effects that had "pops" or distortion.
This fixes almost half of the games sound effects. Because there isn't a good source for these online, the vast majority were fixed manually using Audacity and a lot of time and mental health. The result is magnificent though, so it was worth it. It's one of those things that once you hear it, you can't unhear it. And so many of the games sound effects were affected.