Fix new hole code bottom bias to only affect Link.
I had this thought when working on this, that I only wanted this code to affect Link, yet I forgot to update it so it did only affect Link. The way the code is now, monsters do not fall down the center of the hole due to the bottom bias. This looks somewhat strange especially for Hardhat Beetles that must fall down the hole since they can not be defeated. No matter how many changes or how many times I tweak it, holes will never be perfect. But I can sure keep trying to get as close as I can.