Create a "drowning" animation and rework drowning to be frame based.
The drowning animation in this port was always faked by showing the "swimming" animation for a time then the "diving animation" afterwards. This creates a dedicated "drowning" animation that is 32 frames long like the original games and timing is based on the length of the animation rather than a fixed amount of time. The fake drowning animation showed Link facing the direction he was moving, while the real drowning animation is always supposed to face forward.