Fix jumping out of the water when movement speed is 70% or greater.
To jump out of the water, Link must be at the top of the water and the current swim velocity must be less than the max swim velocity. After 70% bonus move speed, the slightest bit of movement put it over the top making it not possible to jump out of the water. This isn't a perfect fix as after 70% Link will auto jump out of the water, but it's better than not being able to jump at all.