Map Overhaul: Fill in missing details into "black" areas.
I can understand the decision to remove the inaccessible upper layers of maps and just replace them with darkness. The point was probably because of the open camera it would look "weird" if one room rectangle then transitions into nothingness just to have another room rectangle on the other side of that. But I don't care. I want the maps to look as they did in the original game. With classic camera enabled, the missing details really stand out. And when it comes to "modern" vs. "classic", with me "classic" will win out in most cases.