More Hole changes: Accurate offsets, remove offset hack, level 8 map update.
This change alters the way hole reset points work. When Link crosses into a new "field", the reset point is set to the entry point to that field. This meant that the reset point was basically in between two fields. In the original games, the reset point was set so that Link would fully respawn onto the current screen (field). This commit modifies the following:

- Modify the reset point to always have Link within the current field.
- Remove the hack to offset the position when finding doors.
- Fix offset of alternate reset point when direction is set to "1".
- Update level 8 to use alternate reset position on doors that needed hack.