Prevent damage while and shortly after ocarina is played.
Unfortunately the CollisionComponent does not have a way to temporarily deactivate, so this method was implemented instead. This adds a boolean and timer combo that prevents the "HitPlayer" method from firing when Link collides with a damage source (enemy, projectile, etc). For now, it's used solely for when playing the ocarina. When the ocarina is played, "PreventDamage" is set along with a timer that exceeds the time of the song. When the ocarina is done playing, the timer is reduced to 100ms. So in the situation Link is touching a "Pols Voice" enemy and immediately plays the ocarina, only the initial hit will take place. The ocarina plays, damage is negated during this time, the ocarina ends, Link remains invincible for 200ms longer, the Pols Voice dies before this 200ms expires so a second hit does not take place.