Fix quitting/exiting current game to save current game scale.
When setting the scaling value with the controller, the value was never saved. The only place settings were saved is when exiting the settings menu. This commit makes it so quitting or exiting the current game also saves current settings, which in turn saves scaling values set with the trigger buttons.