Reset the dive counter when no longer swimming. Fixes going invincible touching land while diving.
Another epic fail that has probably always been there. The dive counter was the only checked when detecting a hit, effectively making Link invincible if it was above 0. Touching land should have reset the dive counter but it didn't. Or, it would have been enough to also check for swimming state when checking if currently diving. This adds both fixes.