Remove collection cooldown from fairies when collecting with Hookshot/Boomerang.
For some reason, a fairy can not be collected for the first 500ms after it spawns. This can cause a weird bug where if destroying a spark with the boomerang and collecting the fairy immediately sends it flying past Link and into the wall. The next boomerang throw then respawns the fairy and it is collected. So this commit just sets the cooldown to 0 when collected with either hookshot or boomerang.

This also fixes what appears to be an extremely rare bug since I've never seen it. The collection cooldown had to be less than zero in order to collect it. Since the cooldown is subtracted by delta time, chances of landing exactly on zero were really slim. But it was a possibility which would make the fairy uncollectable.