Prevent spawning a "ghost BowWow" after turning him in to Madame MeowMeow.
Efficiently tracking the "BowWow" object proved to be quite a challenge. If one were to try to "find" the object after he was already spawned, it would require loop through all map objects and checking the types. This is not cheap to do, and it's much more efficient to just store the object in a variable when it's created. Problem is, many maps create a BowWow instance even if he's not currently a follower, which tends to spawn a "ghost BowWow". So to prevent this, check the "has_bowWow" state to see if Link has recently saved BowWow and not yet turned him back in, and also don't try to spawn him in dungeons.