Fix Marin-Walrus sequence crash and make compatible with Classic Camera.
For some reason this seems to have started crashing with a "Map is null" error. I've never understood this crash or why it starts happening, and the fix always seems to be to store the map parameter passed on the object's creation to a variable and reference that instead. I don't know if the crash only happened with Classic Camera or if it was some of the ObjMarin changes over time. But at least now it's fixed.

This also updates the walrus sequence overall so it works with Classic Camera. The rabbit's spawn position was moved into the same field rectangle and his despawn point was also slightly shifted. Before it wouldn't even spawn since it was outside of the field rectangle. There is no good way to add it to the exceptions list, so keeping it in the field is the path of least resistance.