Classic Camera: Add more objects to "always animate" group.
The rocks near level 8, the Stone Hinox rocks, the color dungeon boss Hardhit Beetle's projectiles, fireballs, vireballs, and even the lamps inside the egg all need to be animated even when outside of the current field. One thing I find weird is that all objects added to this group need a workaround because the map reference is null. The current map shouldn't change. So when it's null I just reference the map object from ObjLink since it's the same map reference anyway. It's a weird issue but at least the workaround is simple enough.