Apply "always animate" to both standard and classic camera modes.
Even with a much larger viewport, the same objects that benefit from the "always animate" list can also benefit with the standard camera. I was hesitant to apply this system to the standard camera as it served as a fallback if the classic camera was currently broken. But I think it's reached a stable enough point to always animate these objects no matter what camera is used.