Rework how Ocarina Listener component works and shrink Egg rectangle.
The ocarina listener component used to work in that Link plays a song, a fixed size rectangle is checked around him for objects with "listeners", and if both are within that rectangle, the code is ran. All listeners used the same size rectangle as it was a local variable on ObjLink with no way to configure it.

This commit reworks it so that the initial "listener" rectangle is larger, so objects can listen further away, but then another rectangle is created around the object to check if Link is in within it. Before the rectangle was a fixed size, now the rectangle on the listener object can be configured so that the exact size of how far away Link should be can be configured on a per object basis. This was done for the egg in this commit, so that in Classic Camera he has to be within the same field to open the passageway.