Revert b4e77f2 adding Ravens back to "always animate" list.
This has been a confusing situation. The problem seems to not have been the Raven directly, but rather me moving the clear of the always animate list to just before using it rather than after the loop. After that, the Raven issue went away so I think it's safe to add them back to "always animate" objects.