Revamp the "always animate" system to be safer and more reliable.
This is a rather large change to the system that tracks game objects that are added to the always animate list. There are some very crucial changes:

- The main/temp lists are no longer static and exist only on the "Map" object that created the instance of the "ObjectManager".
- Adding, removing, or accessing the always animate lists has been made thread safe using locks.
- A method was created to safely add objects to the list on ObjectManager which automates the process of locking before adding.
- The loop itself still uses a "temporary" list so the main list can be safely modified during an update loop.
- Exception catching was added to the loop in the unlikely scenario that the collection is modified during a loop.
- External component update loops (systemBody, systemAI, systemAnimator) require the object manager instance to be passed to them and are  instantiated on object manager creation.

While I didn't see many problems with the system before this update, I was still experiencing random crashes when using the editor. Mostly with the exception error that the always animate list was modified during the loop. This so far seems to have resolved that issue, and should hopefully also resolve any "map is null" errors now that the lists are bound to the current map.