Store the "previous field" for more than a single update loop.
The "PreviousField" variable was only different than the "CurrentField" for a single loop iteration. It's purpose was only to detect a field change and set the variable "FieldChange" to true, which of course, triggers anything that requires a field change. This commit holds onto the previous field until it is overwritten with a new previous field, and uses a new "ContrastField" to detect a field change. This way there is now a permanent reference to both the current and the previous field Link is/was on.