Fix movable stones in Ukuku Prairie upper cave near the pig moblins.
There is a secret room in a cave behind a destroyable wall with a bunch of Spiny Beetles wearing skulls. The room has a bunch of stones that can be pushed. In this port, all the stones were pushable. In the original game, only some of them are pushable. Match the stones in this port to the stones in the original.