Fix a few issues with the Cucco attack swarm.
When attacking a chicken more than 35 times, it will spawn a swarm of them that attacks. The old logic was that you would have to walk far enough away from the bird to stop the attack wave. This is still true, but if Classic Camera is active you can just do a field change. Also, the hit counter did not reset after the attack wave ended, which it now properly does. Finally, they are now added to the "always animate" list so they don't get stuck at the edges of the screen.