Editor: Create higher definition icons of existing icons.
The icons were upscaled by 4x, but looked pretty meh as a result. I have redrawn all of them to look much nicer and match that 4x resolution since I spend so much time in the editor. Also this colors the "object respawner" orange to help separate it from the "enemy respawner" which uses the same icon.