Classic Camera: Implement Red Zol/Gel field transition respawn logic.
This ended up being a lot more difficult than it should have been. In the original game, if both Gels are alive, the Red Zol is respawned. If one of them is killed, it does not respawn. I knew the solution was to link the enemies together somehow and track the state of each other but I struggled to come up with a good solution for some time.

I first tried to link the Gels to the Zol and keep the Zol "alive" even after death just invisible, but this was a terrible solution and I could not solve all the bugs. The idea was that the Zol was the "master object" that tracked everything, and was responsible for respawning itself. This idea flopped horrendously though.

The best solution ended up being to have the two Gels track each other. When created on the Red Zol init method, both gels are given a link to the other that also assigns a "main" bool and the Red Zol respawn position. The bool is used to track "Gel 0" since we only need one of them to respawn the Zol. When a Gel is killed, the GameObject property "IsDead" is set true to prevent respawn. When a respawn happens, it checks if Gel is "main", both Gels are alive, and both are not inactive. If this passes, respawn the Zol.