Create "no followers" object and add to maps that shouldn't have them.
This creates a new object called "no followers" which can be placed on maps via the editor. This is so the game's codebase does not use hardcoded map names to prevent followers from spawning on them. This commit also adds a a similar check for the "final map" but uses a preexisting object to set the variable "ObjFinalBackground" which is only found on the final stairs.