Allow taking Bow Wow for another walk after dungeon 8.
Like all end game followers, these conditions must be true:
- Level 8 has been completed.
- Nothing is Missable must be enabled.
- No other active followers.

To take Bow Wow for a walk, just walk up to his pole and answer yes to the dialog question. To return Bow Wow, just talk to Madam MeowMeow like you do after the cave rescue. A lot of the functionality for this change is reusing the logic from that event with just a few tweaks.