Rerelease of v1.4.8 via patcher v1.4.8.1.
I made some massive mistakes in the v1.4.8 patcher. Dungeon 3 had a name change and the patcher did not handle it, so no dungeon 3 map was created. This rerelease also contains a few more fixes, and adds the feature to have Bow Wow as a follower after Level 8 is completed and "Nothing is Missable" is activated.