Add "Game" option to auto-select the last selected save.
Whenever the main menu is loaded for the first time, such as after the title screen or when quitting the current game, the last selected save slot will be automatically selected. This is done via "fake" button presses that moves the selection to the save slot. Despite my best efforts, I couldn't do much better than this.