Reference "Owner" Map in ObjectManager rather than ObjLink.Map.
I don't know if it makes a difference since the current map should match both the "Owner" and the map currently stored in "ObjLink", but there probably is a brief moment where the current "Map" object between these two objects diverges. And it's probably more stable to reference "Owner" here since it will always reference the object that is currently accessing the ObjectManager.