Fix bombable walls in some caves + add updated maps with camera field objects.
This fixes the bombable walls of two caves: one near the pig moblins and one on Tal Tal Mountain Range. Visibility was inconsistent as well as the tiles that can give away that they are bombable. Also upload a bunch of map updates that add new camera field objects for the classic camera.