Fix trading Tarin the stick for the honeycomb.
Talking to him would just display "error" instead of actually initiating the trade. I screwed up the script and forgot to start a new dialog path and instead entered the path name. In this situation, it should be "[path:tarin_stick_check]" but I just put "tarin_stick_check" which would try to find a matching dialog that doesn't exist.