Implement new "push button" into first 4 dungeons & color dungeon.
This adds the new "push button" object (no not the kind you push, it's the kind that pushes you) into the first 4 dungeons and the color dungeon. This object is placed in a doorway and pushes the player into the next room. The point is to mimic the behavior of the original game when entering a boss room where Link is forced beyond the door and it slams behind him.