Simulate Pincer falling down hole on brief stun window.
There is a brief window in the original game when it recedes into its hole and the eyes are not visible where magic powder stuns it and it falls down a hole. This enemy is not capable of falling down holes in this port, but we can simulate it instead.