Classic Camera: Raccoon teleport has scrolling effect.
The idea is to make it "feel" more like the teleport in the original game by scrolling the screen during the teleport. The illusion is not as convincing as the same kind of teleport in Level 6 dungeon since the terrain below the screen is much different, but at least the effect is correct and does appear as if the screen is transitioning into the new field.