Add a "Damage Cooldown" modifier.
This adds a new "Modifier" that allows setting the damage cooldown factor which is a value that is multiplied by the damage blink time (66). The default value is 16, which sets damage cooldown to 1056ms, just over a second. The factor can be set anywhere from 0-100, ranging from basically instant death from damage lock to 6600ms of invincibility time.