Rework Level 2 Sword spawn at Seashell Mansion.
This has always had various issues, many have been fixed over time but this is the last one that needed attention. Basically, if the dialog "My work here is finished." was not closed almost immediately, then the timing of the sword spawn music would be completely offset from the actual sword spawn.

One thing to note, is that the sword spawn sound effects is a music track, and it has 2600ms of silence at the start of it. This making timing everything difficult. This commit reworks the logic behind the sword spawn to start the music the moment the dialog is opened. After 2.6 seconds, the dialog will automatically close if the user did not close it, and the sword will go through it's normal spawn routine. No matter how it's approached now, the sword spawn effects and sound effects will always sync up.

To be able to pull this off, I had to create yet another dialog path command "dialog_close", which simply closes the dialog window. AFAIK, there was currently no methods to forcefully close the dialog window, so this adds at least one. Though it's honestly as simple as setting: `Game1.GameManager.InGameOverlay.TextboxOverlay.IsOpen = false;`