Properly clear sword and shield boxes when not in use.
It's uncertain if it caused any bugs other than fairy collection with the lahdmod active, but it's worth fixing regardless. The fairy bug was if collecting the fairy with a sword is enabled, the sword box would hang around after the swing and be able to collect the fairy. Usually the boxes shouldn't do anything when inactive, but there may be other situations I'm not aware of like this one.