Detect climbing to the bottom of a ladder.
There was no good way to go about this. The best I could come up with was to see if the Y position changed between climbing frames. If the position is consistent, then chances are the player hit the bottom so stop climbing. As far as I can tell, this works without any negative side effects.