Allow grabbing and pulling on almost everything.
This replicates the behavior of being able to grab just about everything in the original game and pulling on it, even if it can't be picked up. This probably doesn't match every object 100% as in some new stuff can be pulled, some stuff that could can't be here, but it's general enough that it should be fairly close.