Revert Boomerang body box size and change sword smack method.
The change to the Boomerang body size made it so it could no longer hit dungeon switches from the top so it was reverted back to its 2x2 body size. This would make it hard to ricochet off the sword again, so a new method was implemented. Instead of trying to find contact in the update function, give the boomerang a hittable component and detect sword hits. And instead of using the body box as the reference for this, use the damage box instead. This gives a box of 10x10 to hit and it should be far more responsive.