Add Flame Trap to always animate list.
An exploit I didn't really think of until now, is to adjust the free camera so the Flame Trap is just out of view with Link at the top of the screen. This allows the player to walk past the trap without it ever shooting a single fireball. To prevent this, add the Flame Trap to the always animate list.