Grab bias: go with the absolute highest value possible.
I figured it would probably be safest to go with the highest possible value for the grab bias before it starts to create a gap between Link and the object he is grabbing. The highest value is 0.08 and after that 0.09+ starts to create a gap. This is probably a completely unnecessary change but oh well.