Reset Trendy Game NPC key value pair when entering building.
I assumed that the key would only be checked when it actually changed, not constantly every frame. When starting the game and not letting it finish, the key to enable the NPC is never reset so he remains disabled forever. So when entering the building, reset the key on the NPC's constructor to prevent this from happening.